+++ server/utils/constants.js
@@ -25,6 +25,7 @@ exports.publicEndpoints = [
 	'/v2/scope/scopeAuth',
 	'/v2/accounts/signup',
 	'/v2/oauth/connections/redirect_uri',
+	'/v2/session-logs/insert',
 ];
 
 exports.filterPrefix = "param_";
