+++ server/www/session-logs.js
@@ -1,8 +1,15 @@
 const API = require('../utils/api');
-const sessions = require('../utils/sessions').sessions;
 const commonFun = require('../utils/commonFunctions');
+const atob = require('atob');
 
 class SessionLogs extends API {
+
+	async list() {
+
+		this.user.privilege.needs('user');
+
+		return await this.mysql.query(`SELECT * FROM tb_sessions WHERE account_id = ?`, [this.user.user_id]);
+	};
 
 	async insert() {
 
@@ -11,16 +18,38 @@ class SessionLogs extends API {
 		let params = {
 			user_id: this.request.body.user_id,
 			type: this.request.body.type,
-			description: this.request.body.description,
 			user_agent: this.request.get('user-agent'),
+			expire_time: this.request.body.expire_time,
 			os: user_agent.os,
 			browser: user_agent.browser,
 			refresh_token: this.request.body.refresh_token,
 			ip: this.request.connection.remoteAddress,
 		};
 
-		return await sessions.insert(params);
+		if(params.type != 'login' && params.refresh_token) {
+
+			const check_token = await commonFun.verifyJWT(params.refresh_token);
+
+			if(check_token.error && check_token.message != 'jwt expired')
+				return check_token.message;
+
+			const token_details = JSON.parse(atob(params.refresh_token.split('.')[1]));
+
+			params.user_id = token_details.user_id;
+			params.session_id = token_details.sessionId;
+		}
+
+		delete(params['refresh_token']);
+
+		const result = await this.mysql.query(
+			`INSERT INTO tb_sessions SET ?`,
+			[params],
+			'write'
+		);
+
+		return result;
 	}
 }
 
 exports.insert = SessionLogs;
+exports.sessions = (() => new SessionLogs)();
